Avoid problematic newline escapes in strings.
SBCL does not know how to treat CRLF as #\Newline, so on Windows
systems a checkout of this file with system line endings will fail to
compile, because #\~#\Return#\Newline is not a valid escape sequence.

This could probably also be done with git settings, but this way it
will still compile when software-archeologists from the year 5000 try
to load it in mercurial.