Don't treat connection closes after QUIT as errors.
The server is supposed to send a status line and close the connection
after receiving the QUIT verb. Some servers do this incorrectly and
use a TCP reset, which can result in the status line being discarded
before we can read it; still, we wanted the connection closed, and
it's closed now, so if we see END-OF-FILE just proceed normally.

There is a minor question of whether it breaks API compatibility to
return an empty list in the case of an early EOF. It's probably fine.